refactor CallTaskService with handler classes
