+++ durable-workflow-runtime/src/main/java/com/amannmalik/workflow/runtime/task/call/handler/CallHandler.java
@@ -1,0 +1,7 @@
+package com.amannmalik.workflow.runtime.task.call.handler;
+
+import dev.restate.sdk.WorkflowContext;
+
+public interface CallHandler<T> {
+    void handle(WorkflowContext ctx, T call);
+}
